Reorganize codebase to feature-based structure
