+++ app/src/main/java/com/example/foldtracker/repository/CounterRepository.kt
@@ -1,7 +1,6 @@
 package com.example.foldtracker.repository
 
 import kotlinx.coroutines.flow.Flow
-import java.time.LocalDate
 import androidx.datastore.preferences.core.Preferences
 
 interface CounterRepository {
