--- app/src/main/java/com/example/foldtracker/ui/navigation/AppNavigation.kt
@@ -1,27 +1,0 @@
-package com.example.foldtracker.ui.navigation
-
-import android.annotation.SuppressLint
-import androidx.compose.runtime.Composable
-
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
-import com.example.foldtracker.ui.CounterScreen
-import com.example.foldtracker.ui.StatsScreen
-import com.example.foldtracker.viewmodel.CounterViewModel
-
-@SuppressLint("StateFlowValueCalledInComposition")
-@Composable
-fun AppNavigation(viewModel: CounterViewModel) {
-    val navController = rememberNavController()
-
-    NavHost(navController = navController, startDestination = "counter_screen") {
-        composable("counter_screen") {
-            CounterScreen(viewModel = viewModel,navController)
-        }
-
-        composable("stats_screen"){
-            StatsScreen(viewModel = viewModel)
-        }
-    }
-}
