--- app/src/main/java/com/example/foldtracker/ui/navigation/Screen.kt
@@ -1,7 +1,0 @@
-package com.example.foldtracker.ui.navigation
-
-sealed class Screen(val route:String) {
-    data object Counter : Screen("counter")
-    data object Stats : Screen("stats")
-    data object Achievements : Screen("achievements")
-}
