--- app/src/main/java/com/example/foldtracker/widget/WidgetRefreshCallback.kt
@@ -1,19 +1,0 @@
-package com.example.foldtracker.widget
-
-import android.content.Context
-import android.util.Log
-import androidx.glance.GlanceId
-import androidx.glance.action.ActionParameters
-import androidx.glance.appwidget.action.ActionCallback
-
-class WidgetRefreshCallback : ActionCallback {
-    
-    override suspend fun onAction(
-        context: Context,
-        glanceId: GlanceId,
-        parameters: ActionParameters
-    ) {
-        Log.d("WidgetRefresh", "Refresh button clicked. Updating widget.")
-        FoldCountWidget.updateWidget(context)
-    }
-}
